+++ src/Contracts/SimpleAccessLog.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace aliirfaan\LaravelSimpleAccessLog\Contracts;
+
+interface SimpleAccessLog
+{
+}