+++ src/SimpleAccessLogProvider.php
@@ -4,6 +4,7 @@ namespace aliirfaan\LaravelSimpleAccessLog;
 
 use Illuminate\Support\ServiceProvider;
 use aliirfaan\LaravelSimpleAccessLog\Providers\EventServiceProvider;
+use aliirfaan\LaravelSimpleAccessLog\Contracts\SimpleAccessLog as SimpleAccessLogContract;
 
 class SimpleAccessLogProvider extends ServiceProvider
 {
@@ -29,5 +30,18 @@ class SimpleAccessLogProvider extends ServiceProvider
         $this->publishes([
             __DIR__.'/../config/simple-access-log.php' => config_path('simple-access-log.php'),
         ]);
+
+        $this->registerModelBindings();
+    }
+
+    protected function registerModelBindings()
+    {
+        $config = $this->app->config['simple-access-log'];
+
+        if (! $config) {
+            return;
+        }
+
+        $this->app->bind(SimpleAccessLogContract::class,  $config['access_log_model']);
     }
 }